refactor(vapor): reorg node op helpers + remove children helper
