+++ packages/runtime-vapor/src/dom/template.ts
@@ -12,33 +12,3 @@ export function template(html: string, root?: boolean) {
     return ret
   }
 }
-
-/*! #__NO_SIDE_EFFECTS__ */
-export function children(node: Node, ...paths: number[]): Node {
-  for (const idx of paths) {
-    // In various situations, select the quickest approach.
-    // See https://github.com/vuejs/vue-vapor/pull/263
-    node =
-      idx === 0
-        ? node.firstChild!
-        : idx === 1
-          ? node.firstChild!.nextSibling!
-          : node.childNodes[idx]
-  }
-  return node
-}
-
-export function child(node: ParentNode): Node {
-  return node.firstChild!
-}
-
-export function next(node: Node): Node {
-  return node.nextSibling!
-}
-
-export function nextn(node: Node, offset: number = 1): Node {
-  for (let i = 0; i < offset; i++) {
-    node = node.nextSibling!
-  }
-  return node
-}
